Add Manager integration tests and make Manager.examine work.
git-svn-id: https://fngtps.com/svn/kari/trunk/src@334 7a99d2f0-d533-0410-96fe-f25cb2755c1e

+++ app/models/karidoc_generator.rb
@@ -64,7 +64,7 @@ class KaridocGenerator
   end
   
   def self.clear_if_empty(dir_name)
-    if (Dir.entries(dir_name) - %w(. ..)).empty?
+    if File.exist?(dir_name) and (Dir.entries(dir_name) - %w(. ..)).empty?
       FileUtils.rm_rf(dir_name)
       clear_if_empty(File.dirname(dir_name))
     end
